refactor: move isRef
