+++ packages/react-dnd/src/createSourceConnector.ts
@@ -2,7 +2,7 @@ declare var require: any
 import * as React from 'react'
 import wrapConnectorHooks from './wrapConnectorHooks'
 import { Backend, Unsubscribe, Identifier } from 'dnd-core'
-import { isRef } from './hooks/util'
+import { isRef } from './utils/isRef'
 import { DragSourceOptions, DragPreviewOptions } from './interfaces'
 const shallowEqual = require('shallowequal')
