+++ packages/react-dnd/src/createTargetConnector.ts
@@ -2,7 +2,7 @@ declare var require: any
 import * as React from 'react'
 import wrapConnectorHooks from './wrapConnectorHooks'
 import { Backend, Unsubscribe, Identifier } from 'dnd-core'
-import { isRef } from './hooks/util'
+import { isRef } from './utils/isRef'
 const shallowEqual = require('shallowequal')
 
 export default function createTargetConnector(backend: Backend) {
